+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -10,7 +10,6 @@ import org.slf4j.LoggerFactory;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
 import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspViews;
 
 @Controller
 public class LoginController {
