+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -8,7 +8,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
 import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.view.JspViews;
 
 @Controller
 public class RegisterController {
